requirements.txt
anthropic>=0.40.0
streamlit>=1.35.0
pydantic>=2.7.0
python-dotenv>=1.0.0
pandas>=2.0.0
plotly>=5.20.0
faker>=25.0.0
